Add support for viewing the Hijri date
I used to support viewing the Hijri date in Home Assistant using the Python script that converts the current Gregorian date to Hijri, and a command line sensor in the configuration.yaml that shows the Hijri date...along with automation that updates the date every day at midnight. It works properly for me, and I propose adding this support to the integration, along with updating the date each day. This should work locally (without the need for the Internet).
I also recommend that the prayer times calculation work without a cloud connection because the entities become unavailable when the Home Assistant disconnects from the integration cloud.